celery_rabbitmq/pyproject.toml
[project]
name = "celery-rabbitmq"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "celery>=5.5.3",
    "psycopg2-binary>=2.9.10",
    "python-dotenv>=1.0.0",
    "sqlalchemy>=2.0.43",
    "python-dotenv>=1.0.0",
]
